+++ src/lib.mli
@@ -37,16 +37,7 @@ module Set : Set.S with type elt = t
 
 module Map : Map.S with type key = t
 
-module Status : sig
-  type t =
-    | Installed
-    | Public  of Package.t
-    | Private of Dune_project.Name.t
-
-  val pp : t Fmt.t
-end
-
-val status : t -> Status.t
+val status : t -> Lib_info.Status.t
 
 val package : t -> Package.Name.t option
 
@@ -85,48 +76,6 @@ module Lib_and_module : sig
 
   val link_flags : t list -> mode:Mode.t -> stdlib_dir:Path.t -> _ Arg_spec.t
 
-end
-
-(** {1 Raw library descriptions} *)
-
-(** Information about a library *)
-module Info : sig
-  module Deps : sig
-    type t =
-      | Simple  of (Loc.t * Lib_name.t) list
-      | Complex of Dune_file.Lib_dep.t list
-  end
-
-  (** Raw description of a library, where dependencies are not
-      resolved. *)
-  type t =
-    { loc              : Loc.t
-    ; kind             : Dune_file.Library.Kind.t
-    ; status           : Status.t
-    ; src_dir          : Path.t
-    ; obj_dir          : Path.t
-    ; version          : string option
-    ; synopsis         : string option
-    ; archives         : Path.t list Mode.Dict.t
-    ; plugins          : Path.t list Mode.Dict.t
-    ; foreign_archives : Path.t list Mode.Dict.t (** [.a/.lib/...] files *)
-    ; jsoo_runtime     : Path.t list
-    ; requires         : Deps.t
-    ; ppx_runtime_deps : (Loc.t * Lib_name.t) list
-    ; pps              : (Loc.t * Dune_file.Pp.t) list
-    ; optional         : bool
-    ; virtual_deps     : (Loc.t * Lib_name.t) list
-    ; dune_version : Syntax.Version.t option
-    ; sub_systems      : Dune_file.Sub_system_info.t Sub_system_name.Map.t
-    }
-
-  val of_library_stanza
-    : dir:Path.t
-    -> ext_lib:string
-    -> Dune_file.Library.t
-    -> t
-
-  val of_findlib_package : Findlib.Package.t -> t
 end
 
 (** {1 Errors} *)
@@ -250,8 +199,8 @@ module DB : sig
   module Resolve_result : sig
     type nonrec t =
       | Not_found
-      | Found    of Info.t
-      | Hidden   of Info.t * string
+      | Found    of Lib_info.t
+      | Hidden   of Lib_info.t * string
       | Redirect of t option * Lib_name.t
   end