+++ test/blackbox-tests/test-cases/dune-ppx-driver-system/run.t
@@ -2,8 +2,8 @@ No ppx driver found
 
   $ dune build foo1.cma
   File "dune", line 6, characters 13-18:
-   (preprocess (pps)))
-               ^^^^^
+  6 |  (preprocess (pps)))
+                   ^^^^^
   Error: You must specify at least one ppx rewriter.
   [1]
 
@@ -11,8 +11,8 @@ Too many drivers
 
   $ dune build foo2.cma
   File "dune", line 13, characters 13-28:
-   (preprocess (pps ppx1 ppx2)))
-               ^^^^^^^^^^^^^^^
+  13 |  (preprocess (pps ppx1 ppx2)))
+                    ^^^^^^^^^^^^^^^
   Error: Too many incompatible ppx drivers were found: foo.driver2 and
   foo.driver1.
   [1]
@@ -21,8 +21,8 @@ Not compatible with Dune
 
   $ dune build foo3.cma
   File "dune", line 20, characters 13-28:
-   (preprocess (pps ppx_other)))
-               ^^^^^^^^^^^^^^^
+  20 |  (preprocess (pps ppx_other)))
+                    ^^^^^^^^^^^^^^^
   Error: No ppx driver were found. It seems that ppx_other is not compatible
   with Dune. Examples of ppx rewriters that are compatible with Dune are ones
   using ocaml-migrate-parsetree, ppxlib or ppx_driver.
@@ -69,7 +69,7 @@ Test that going throught the -ppx option of the compiler works
   $ dune build test_ppx_staged.cma
       ocamldep .test_ppx_staged.objs/test_ppx_staged.ml.d
   tool name: ocamldep
-  args:--cookie library-name="test_ppx_staged"
+  args:--as-ppx --cookie library-name="test_ppx_staged"
         ocamlc .test_ppx_staged.objs/test_ppx_staged.{cmi,cmo,cmt}
   tool name: ocamlc
-  args:--cookie library-name="test_ppx_staged"
+  args:--as-ppx --cookie library-name="test_ppx_staged"