+++ test/blackbox-tests/test-cases/misc/run.t
@@ -1,7 +1,7 @@
   $ dune runtest --display short
   File "dune", line 44, characters 19-42:
-   (deps (glob_files dir-that-doesnt-exist/*)))
-                     ^^^^^^^^^^^^^^^^^^^^^^^
+  44 |  (deps (glob_files dir-that-doesnt-exist/*)))
+                          ^^^^^^^^^^^^^^^^^^^^^^^
   Warning: Directory dir-that-doesnt-exist doesn't exist.
           diff alias runtest
           diff alias runtest