+++ test/blackbox-tests/test-cases/promote/run.t
@@ -1,3 +1,6 @@
+General tests
+--------------------------
+
   $ printf titi > x
 
   $ dune build --display short --diff-command false @blah 2>&1 | sed 's/.*false.*/DIFF/'
@@ -28,3 +31,29 @@ Otherwise this test fails on OSX
   $ dune build --display short --diff-command false @blah
   $ cat x
   toto
+
+Test single file promotion
+--------------------------
+
+  $ printf a > x
+  $ printf a > y
+  $ dune build --display short --diff-command false @blah @blah2 2>&1 | sed 's/.*false.*/DIFF/'
+            sh (internal) (exit 1)
+  DIFF
+            sh (internal) (exit 1)
+  DIFF
+  $ dune promote x
+  Promoting _build/default/x.gen to x.
+  $ cat x
+  toto
+  $ cat y
+  a
+  $ dune promote y
+  Promoting _build/default/y.gen to y.
+  $ cat x
+  toto
+  $ cat y
+  titi
+  $ dune promote x y
+  Warning: Nothing to promote for x.
+  Warning: Nothing to promote for y.