+++ test/blackbox-tests/test-cases/several-packages/run.t
@@ -5,9 +5,9 @@ displayed. This can happen for:
 
   $ dune build --root executable
   File "dune", line 1, characters 0-43:
-  1: (executable
-  2:   (public_name an_executable)
-  3: )
+  1 | (executable
+  2 |   (public_name an_executable)
+  3 | )
   Error: I can't determine automatically which package this stanza is for.
   I have the choice between these ones:
   - pkg1 (because of pkg1.opam)
@@ -19,8 +19,8 @@ displayed. This can happen for:
 
   $ dune build --root documentation
   File "dune", line 1, characters 0-15:
-  (documentation)
-  ^^^^^^^^^^^^^^^
+  1 | (documentation)
+      ^^^^^^^^^^^^^^^
   Error: I can't determine automatically which package this stanza is for.
   I have the choice between these ones:
   - pkg1 (because of pkg1.opam)
@@ -32,10 +32,10 @@ displayed. This can happen for:
 
   $ dune build --root install
   File "dune", line 1, characters 0-44:
-  1: (install
-  2:  (section etc)
-  3:  (files file.conf)
-  4: )
+  1 | (install
+  2 |  (section etc)
+  3 |  (files file.conf)
+  4 | )
   Error: I can't determine automatically which package this stanza is for.
   I have the choice between these ones:
   - pkg1 (because of pkg1.opam)